pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eigsep_sensors"
version = "0.0.1"
description = "Software to control EIGSEP sensors."
readme = "README.md"
license = {text = "MIT"}
authors = [
  { name="Pranav Karthik", email="pranavkarthik@berkeley.edu" },
  { name="Richard Saeed", email="richardjs@berkeley.edu" },
  { name="Christian Hellum Bye", email="cbh@berkeley.edu" },
]
requires-python = ">=3.9"
classifiers = [
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
  "numpy",
  "mpremote",
  "smbus2",
  "adafruit-circuitpython-mma8451",
  "adafruit-circuitpython-bno08x",
  "adafruit-blinka<8.60.0",
  "board",
  "mockserial @ git+https://github.com/christianhbye/mockserial.git"
]

[project.optional-dependencies]
dev = [
  "black",
  "flake8",
  "pytest",
  "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
  "--cov=src",
  "--cov-report=term-missing",
  "--cov-report=xml",
  "--junitxml=junit.xml",
  "-o junit_family=legacy",
]

[tool.black]
line-length = 79
